pyproject.toml
[project]
name = "llmchat"
dependencies = ["logger", "openai", "numpy", "PyAudio", "pynacl", "requests", "typing", "discord.py @ git+https://github.com/Sheepposu/discord.py@voice-receive", "SpeechRecognition", "Pillow", "tiktoken", "scipy", "poe-api", "anthropic"]
version = "1.0.0"
scripts = {llmchat = "llmchat.cli:main"}

[project.optional-dependencies]
blip = ["transformers[torch]", "torchaudio"]
llama = ["langchain", "llama-cpp-python"]
voice = ["elevenlabs", "suno-bark @ git+https://github.com/suno-ai/bark.git", "azure-cognitiveservices-speech", "transformers[torch]"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[options.packages.find]
where = ""

[tool.setuptools.package-data]
"llmchat" = ["config.example.ini"]

[tool.setuptools]
packages = ["llmchat", "llmchat.llm_sources", "llmchat.sr_sources", "llmchat.tts_sources"]